+++ .settings/org.eclipse.wst.common.component
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="UTF-8"?><project-modules id="moduleCoreId" project-version="1.5.0">
-                                    
+                                        
     
     
     
@@ -7,8 +7,10 @@
     
     
     
-    <wb-module deploy-name="Acme-Toolkits-22.1">
-                                                                        
+    
+    <wb-module deploy-name="Acme-Toolkits-22.1-22.1">
+                                                                                
+        
         
         
         
@@ -17,7 +19,8 @@
         
         
         <wb-resource deploy-path="/" source-path="/target/m2e-wtp/web-resources"/>
-                                                                        
+                                                                                
+        
         
         
         
@@ -26,7 +29,8 @@
         
         
         <wb-resource deploy-path="/" source-path="/src/main/webapp" tag="defaultRootSource"/>
-                                                                        
+                                                                                
+        
         
         
         
@@ -35,7 +39,8 @@
         
         
         <wb-resource deploy-path="/WEB-INF/classes" source-path="/src/main/java"/>
-                                                                        
+                                                                                
+        
         
         
         
@@ -44,7 +49,8 @@
         
         
         <wb-resource deploy-path="/WEB-INF/classes" source-path="/src/main/resources"/>
-                                                                        
+                                                                                
+        
         
         
         
@@ -53,21 +59,25 @@
         
         
         <wb-resource deploy-path="/WEB-INF/classes" source-path="/acme-framework"/>
-                                        
+                                                
+        
         
         
         
         <wb-resource deploy-path="/WEB-INF/classes" source-path="/target/generated-sources/annotations"/>
-                                        
+                                                
+        
         
         
         
         <wb-resource deploy-path="/WEB-INF/classes" source-path="/target/generated-test-sources/test-annotations"/>
-                                
+                                        
+        
         
         
         <wb-resource deploy-path="/WEB-INF/classes" source-path="/.apt_generated"/>
-                                                                        
+                                                                                
+        
         
         
         
@@ -76,8 +86,9 @@
         
         
         <property name="java-output-path" value="/Acme-Toolkits/target/classes"/>
-        <property name="context-root" value="Acme-Toolkits"/>
-                                                                    
+        <property name="context-root" value="Acme-Toolkits-22.1"/>
+                                                                            
+    
     
     
     
@@ -86,7 +97,8 @@
     
     
     </wb-module>
-                                
+                                    
+
