+++ src/main/java/acme/entities/chirps/Chirp.java
@@ -1,46 +1,47 @@
-package acme.entities.patronageReport;
+package acme.entities.chirps;
 
 import java.util.Date;
 
 import javax.persistence.Entity;
 import javax.persistence.Temporal;
 import javax.persistence.TemporalType;
+import javax.validation.constraints.Email;
 import javax.validation.constraints.NotBlank;
-import javax.validation.constraints.Past;
-import javax.validation.constraints.Pattern;
 
 import org.hibernate.validator.constraints.Length;
-import org.hibernate.validator.constraints.URL;
 
 import acme.framework.entities.AbstractEntity;
 import lombok.Getter;
 import lombok.Setter;
 
+
 @Entity
 @Getter
 @Setter
-public class PatronageReport extends AbstractEntity{
+public class Chirp extends AbstractEntity {
 	
 	// Serialisation identifier -----------------------------------------------
 
-	protected static final long	serialVersionUID	= 1L;
+	protected static final long		serialVersionUID	= 1L;
 
 	// Attributes -------------------------------------------------------------
 
-	@NotBlank
-	@Pattern(regexp = "^[A-Z]{3}-[0-9]{3}(-[A-Z])?:[0-9]{4}$")
-	protected String sequenceNumber;
-	
-	@Past
 	@Temporal(TemporalType.TIMESTAMP)
-	protected Date creationTime;
+	protected Date creationMoment;
+	
+	@NotBlank
+	@Length(max=100)
+	protected String title;
+	
+	@NotBlank
+	@Length(max=100)
+	protected String author;
 	
 	@NotBlank
 	@Length(max=255)
-	protected String memorandum;
+	protected String body;
 	
-	@URL
-	protected String info;
-	
+	@Email
+	protected String email;
 	
 }
