+++ src/main/java/acme/entities/items/ItemType.java
@@ -1,4 +1,4 @@
-package acme.entities.item;
+package acme.entities.items;
 
 
 public enum ItemType {
