+++ src/main/java/acme/entities/patronagereports/PatronageReport.java
@@ -1,47 +1,46 @@
-package acme.entities.chirp;
+package acme.entities.patronagereports;
 
 import java.util.Date;
 
 import javax.persistence.Entity;
 import javax.persistence.Temporal;
 import javax.persistence.TemporalType;
-import javax.validation.constraints.Email;
 import javax.validation.constraints.NotBlank;
+import javax.validation.constraints.Past;
+import javax.validation.constraints.Pattern;
 
 import org.hibernate.validator.constraints.Length;
+import org.hibernate.validator.constraints.URL;
 
 import acme.framework.entities.AbstractEntity;
 import lombok.Getter;
 import lombok.Setter;
 
-
 @Entity
 @Getter
 @Setter
-public class Chirp extends AbstractEntity {
+public class PatronageReport extends AbstractEntity{
 	
 	// Serialisation identifier -----------------------------------------------
 
-	protected static final long		serialVersionUID	= 1L;
+	protected static final long	serialVersionUID	= 1L;
 
 	// Attributes -------------------------------------------------------------
 
+	@NotBlank
+	@Pattern(regexp = "^[A-Z]{3}-[0-9]{3}(-[A-Z])?:[0-9]{4}$")
+	protected String sequenceNumber;
+	
+	@Past
 	@Temporal(TemporalType.TIMESTAMP)
-	protected Date creationMoment;
-	
-	@NotBlank
-	@Length(max=100)
-	protected String title;
-	
-	@NotBlank
-	@Length(max=100)
-	protected String author;
+	protected Date creationTime;
 	
 	@NotBlank
 	@Length(max=255)
-	protected String body;
+	protected String memorandum;
 	
-	@Email
-	protected String email;
+	@URL
+	protected String info;
+	
 	
 }
