+++ src/main/java/acme/entities/patronages/Patronage.java
@@ -17,7 +17,7 @@ import javax.validation.constraints.Pattern;
 import org.hibernate.validator.constraints.Length;
 import org.hibernate.validator.constraints.URL;
 
-import acme.entities.patronageReport.PatronageReport;
+import acme.entities.patronagereports.PatronageReport;
 import acme.framework.datatypes.Money;
 import acme.framework.entities.AbstractEntity;
 import acme.roles.Inventor;
@@ -63,7 +63,6 @@ public class Patronage  extends AbstractEntity{
 	@NotNull
 	@Temporal(TemporalType.TIMESTAMP)
 	protected Date						endDate;
-	
 	
 	@URL
 	protected String					info;
