+++ src/main/java/acme/entities/toolkits/Toolkits.java
@@ -2,11 +2,11 @@ package acme.entities.toolkits;
 
 import javax.persistence.Column;
 import javax.persistence.Entity;
-import javax.validation.constraints.Max;
 import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.NotNull;
 import javax.validation.constraints.Pattern;
 
+import org.hibernate.validator.constraints.Length;
 import org.hibernate.validator.constraints.URL;
 
 import acme.framework.entities.AbstractEntity;
@@ -41,15 +41,15 @@ public class Toolkits extends AbstractEntity {
 	protected String code;
 	
 	@NotBlank
-	@Max(101)
+	@Length(max=100)
 	protected String title;
 	
 	@NotBlank
-	@Max(256)
+	@Length(max=255)
 	protected String description;
 	
 	@NotBlank
-	@Max(256)
+	@Length(max=255)
 	protected String assemblyNotes;
 	
 	@URL
